+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Updated included schema files, and added instructions on how to do that to the README
+- Refactored libcovebods/lib/common_checks.py
+  - Classes RunAdditionalChecks and  GetStatistics are gone, refactored to LegacyChecks and LegacyStatistics
+  - New abstract class AdditionalCheck - can be extended by other classes to have better seperation in the future 
+  - New function process_additional_checks available to do work
 
 ## [0.12.0] - 2021-10-26